Minor renames and refactorings
