+++ src/main/kotlin/com/sourcegraph/cody/agent/CodyAgentService.kt
@@ -12,7 +12,7 @@ import com.sourcegraph.cody.chat.AgentChatSessionService
 import com.sourcegraph.cody.config.CodyApplicationSettings
 import com.sourcegraph.cody.context.RemoteRepoSearcher
 import com.sourcegraph.cody.edit.EditService
-import com.sourcegraph.cody.edit.FixupService
+import com.sourcegraph.cody.edit.LensesService
 import com.sourcegraph.cody.error.CodyConsole
 import com.sourcegraph.cody.ignore.IgnoreOracle
 import com.sourcegraph.cody.listeners.CodyFileEditorListener
@@ -83,7 +83,7 @@ class CodyAgentService(private val project: Project) : Disposable {
       }
 
       agent.client.onCodeLensesDisplay = Consumer { params ->
-        FixupService.getInstance(project).updateLenses(params.uri, params.codeLenses)
+        LensesService.getInstance(project).updateLenses(params.uri, params.codeLenses)
       }
 
       agent.client.onTextDocumentEdit = Function { params ->
