+++ src/main/kotlin/com/sourcegraph/cody/edit/actions/lenses/EditRetryAction.kt
@@ -2,12 +2,12 @@ package com.sourcegraph.cody.edit.actions.lenses
 
 import com.intellij.openapi.application.runInEdt
 import com.sourcegraph.cody.edit.EditCommandPrompt
-import com.sourcegraph.cody.edit.FixupService
+import com.sourcegraph.cody.edit.actions.EditCodeAction
 
 class EditRetryAction :
     LensEditAction({ project, _, editor, taskId ->
       runInEdt {
-        val completedFixup = FixupService.getInstance(project).completedFixups[taskId.id]
+        val completedFixup = EditCodeAction.completedEditTasks[taskId.id]
         if (completedFixup != null) {
           runInEdt {
             EditCommandPrompt(project, editor, "Edit instructions and Retry", completedFixup)
