+++ src/main/kotlin/com/sourcegraph/cody/edit/actions/lenses/EditShowDiffAction.kt
@@ -3,7 +3,7 @@ package com.sourcegraph.cody.edit.actions.lenses
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.EditorFactory
 import com.sourcegraph.cody.agent.CodyAgentService
-import com.sourcegraph.cody.edit.actions.ShowDocumentDiffAction
+import com.sourcegraph.common.ShowDocumentDiffAction
 
 class EditShowDiffAction :
     LensEditAction({ project, event, editor, taskId ->
