+++ src/main/kotlin/com/sourcegraph/cody/edit/widget/LensLabel.kt
@@ -3,7 +3,7 @@ package com.sourcegraph.cody.edit.widget
 import com.intellij.openapi.editor.event.EditorMouseEvent
 import com.intellij.ui.JBColor
 import com.intellij.util.ui.UIUtil
-import com.sourcegraph.cody.edit.FixupService.Companion.SEPARATOR
+import com.sourcegraph.cody.edit.LensesService.Companion.SEPARATOR
 import java.awt.FontMetrics
 import java.awt.Graphics2D
